implement card-game stage
